feat: aggregate task feedback service
