+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/feedback/dto/RetrieveTaskFeedbackResult.java
@@ -2,8 +2,8 @@ package com.LetMeDoWith.LetMeDoWith.application.feedback.dto;
 
 import com.LetMeDoWith.LetMeDoWith.domain.feedback.repository.dto.DowithTaskFeedbackQueryDto;
 import com.LetMeDoWith.LetMeDoWith.domain.feedback.repository.dto.TaskFeedbackTemplateQueryDto;
-import com.LetMeDoWith.LetMeDoWith.domain.task.enums.CountryCode;
 import io.swagger.v3.oas.annotations.media.Schema;
+
 import java.util.List;
 
 @Schema(description = "두윗모드 Task 잔소리 조회 결과")
@@ -36,19 +36,8 @@ public record RetrieveTaskFeedbackResult(Long totalCount, List<TaskFeedbackDto> 
             @Schema(description = "잔소리 보낸사람 ID", example = "(TSID)") String senderId,
             @Schema(description = "잔소리 받는사람 닉네임", example = "feedbackSender123") String senderNickname,
             @Schema(description = "잔소리 받는사람 프로필 이미지 URL", example = "https://example.com/profile.jpg")
-                    String senderProfileImageUrl,
+            String senderProfileImageUrl,
             @Schema(description = "잔소리 확인여부", example = "false") Boolean isChecked,
-            @Schema(description = "잔소리 템플릿") TaskFeedbackTemplateDto taskFeedbackTemplate) {}
-
-    public record TaskFeedbackTemplateDto(
-            @Schema(description = "잔소리 템플릿 ID", example = "1") Long id,
-            @Schema(description = "잔소리 템플릿 언어", example = "ko") CountryCode language,
-            @Schema(description = "잔소리 템플릿 메시지", example = "잔소리 템플릿 메시지") String message,
-            @Schema(description = "잔소리 템플릿 이모지 URL", example = "https://example.com/emoji.png") String emojiUrl) {
-
-        public static TaskFeedbackTemplateDto from(TaskFeedbackTemplateQueryDto template) {
-            return new TaskFeedbackTemplateDto(
-                    template.id(), template.language(), template.message(), template.emojiUrl());
-        }
+            @Schema(description = "잔소리 템플릿") TaskFeedbackTemplateDto taskFeedbackTemplate) {
     }
 }
