+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/feedback/dto/RetrieveTaskFeedbackTemplatesResult.java
@@ -1,8 +1,8 @@
 package com.LetMeDoWith.LetMeDoWith.application.feedback.dto;
 
 import com.LetMeDoWith.LetMeDoWith.domain.feedback.repository.dto.TaskFeedbackTemplateQueryDto;
-import com.LetMeDoWith.LetMeDoWith.domain.task.enums.CountryCode;
 import io.swagger.v3.oas.annotations.media.Schema;
+
 import java.util.List;
 
 @Schema(description = "잔소리 템플릿 조회 결과")
@@ -14,10 +14,4 @@ public record RetrieveTaskFeedbackTemplatesResult(List<TaskFeedbackTemplateDto> 
                 .toList();
         return new RetrieveTaskFeedbackTemplatesResult(dtos);
     }
-
-    public record TaskFeedbackTemplateDto(
-            @Schema(description = "잔소리 템플릿 ID", example = "1") Long id,
-            @Schema(description = "잔소리 언어", example = "KR") CountryCode language,
-            @Schema(description = "잔소리 메시지", example = "오늘도 열심히 하셨나요?") String message,
-            @Schema(description = "잔소리 이모지 URL", example = "https://example.com/emoji.png") String emojiUrl) {}
 }
