+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/feedback/repository/dto/AggregateTaskFeedbacksQueryDto.java
@@ -1,0 +1,4 @@
+package com.LetMeDoWith.LetMeDoWith.domain.feedback.repository.dto;
+
+public record AggregateTaskFeedbacksQueryDto() {
+}
